requirements.txt
flask
Flask-Sock
python-dotenv
requests
google-generativeai